+++ src/components/auctions/controller.js
@@ -1,6 +1,6 @@
 import { createAuction } from './service';
 import { isLotTaken, updateLot } from '../lots/service';
-import lotStatuses from '../../enums/lotStatuses';
+import LOT_STATUSES from '../../constants/lotStatuses';
 
 export const create = async (req, res, next) => {
   const {
@@ -14,13 +14,14 @@ export const create = async (req, res, next) => {
       });
     }
 
-    const auction = await createAuction({
-      minimumAcceptablePrice,
-      minimumStep,
-      lot: lotId,
-    });
-
-    await updateLot(lotId, { status: lotStatuses.IN_SALE });
+    const [auction] = await Promise.all([
+      createAuction({
+        minimumAcceptablePrice,
+        minimumStep,
+        lot: lotId,
+      }),
+      updateLot(lotId, { status: LOT_STATUSES.IN_SALE }),
+    ]);
 
     return res.json(auction);
   } catch (e) {
