+++ src/components/lots/model.js
@@ -1,6 +1,6 @@
 import { Schema, model } from 'mongoose';
 import { model as UserModel } from '@app/components/users';
-import lotStatuses from '../../enums/lotStatuses';
+import LOT_STATUSES from '../../constants/lotStatuses';
 
 const LotSchema = new Schema(
   {
@@ -19,8 +19,8 @@ const LotSchema = new Schema(
     },
     status: {
       type: String,
-      enum: Object.values(lotStatuses),
-      default: lotStatuses.AVAILABLE,
+      enum: Object.values(LOT_STATUSES),
+      default: LOT_STATUSES.AVAILABLE,
     },
   },
   {
