+++ src/constants/auctionStatuses.js
