+++ src/constants/lotStatuses.js
